requirements.txt
# Core Google Cloud dependencies
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.13.1
google-cloud-functions-framework==3.5.0
google-cloud-monitoring==2.16.0
google-cloud-logging==3.8.0

# Vertex AI and ML
vertexai==1.38.1

# Graph database
neo4j==5.14.1

# Web framework and API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data processing and utilities
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
python-dateutil==2.8.2

# Async processing
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Additional utilities for real-time processing
concurrent-futures==3.1.1
threading-timer==0.1.0

# For external integrations (optional)
twilio==8.10.0
sendgrid==6.10.0
slack-sdk==3.23.0
